app/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
aiofiles==23.2.1
python-magic==0.4.27
moviepy==1.0.3
slowapi==0.1.9
